+++ Resources/Prototypes/Entities/Effects/chemistry_effects.yml
@@ -77,7 +77,6 @@
       state: foam-north
     - map: [ "enum.EdgeLayer.West" ]
       state: foam-west
-  - type: SmoothEdge
   - type: IconSmooth
     mode: NoSprite
   - type: FoamVisuals
@@ -165,7 +164,6 @@
         - WallLayer
   - type: Occluder
   - type: Appearance
-  - type: SmoothEdge
   - type: IconSmooth
     key: walls
     mode: NoSprite
